fixed error list handling workspace only item
also added test for those cases
