+++ src/VisualStudio/Core/Def/Implementation/TableDataSource/VisualStudioBaseDiagnosticListTable.LiveTableDataSource.cs
@@ -276,13 +276,13 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation.TableDataSource
                     {
                         case StandardTableKeyNames.ErrorRank:
                             content = ValueTypeCache.GetOrCreate(GetErrorRank(data));
-                            return true;
+                            return content != null;
                         case StandardTableKeyNames.ErrorSeverity:
                             content = ValueTypeCache.GetOrCreate(GetErrorCategory(data.Severity));
-                            return true;
+                            return content != null;
                         case StandardTableKeyNames.ErrorCode:
                             content = data.Id;
-                            return true;
+                            return content != null;
                         case StandardTableKeyNames.ErrorCodeToolTip:
                             content = GetHelpLinkToolTipText(data);
                             return content != null;
@@ -291,19 +291,19 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation.TableDataSource
                             return content != null;
                         case StandardTableKeyNames.ErrorCategory:
                             content = data.Category;
-                            return true;
+                            return content != null;
                         case StandardTableKeyNames.ErrorSource:
                             content = ValueTypeCache.GetOrCreate(GetErrorSource(_source.BuildTool));
-                            return true;
+                            return content != null;
                         case StandardTableKeyNames.BuildTool:
                             content = GetBuildTool(_source.BuildTool);
                             return content != null;
                         case StandardTableKeyNames.Text:
                             content = data.Message;
-                            return true;
+                            return content != null;
                         case StandardTableKeyNames.DocumentName:
                             content = GetFileName(data.DataLocation?.OriginalFilePath, data.DataLocation?.MappedFilePath);
-                            return true;
+                            return content != null;
                         case StandardTableKeyNames.Line:
                             content = data.DataLocation?.MappedStartLine ?? 0;
                             return true;
